refactor: Mudando chamada autenticação para o MapsFragment por ViewModels
