+++ app/src/main/java/com/example/rastreioonibus/di/androdModule.kt
@@ -1,13 +1,13 @@
 package com.example.rastreioonibus.di
 
 import com.example.rastreioonibus.utils.API
-import com.example.rastreioonibus.http.HttpUtils
+import com.example.rastreioonibus.http.HttpRepository
 import com.example.rastreioonibus.http.RastreioOnibusApi
+import com.example.rastreioonibus.mapsInicio.MapsViewModel
 import okhttp3.OkHttpClient
 import okhttp3.logging.HttpLoggingInterceptor
 import org.koin.dsl.module
 import com.google.gson.GsonBuilder
-import org.koin.android.ext.koin.androidApplication
 import retrofit2.converter.gson.GsonConverterFactory
 import retrofit2.Retrofit
 
@@ -15,7 +15,11 @@ val androidModule = module {
     single { this }
 
     factory {
-        HttpUtils(api = get())
+        HttpRepository(api = get())
+    }
+
+    factory {
+        MapsViewModel(repo = get())
     }
 
     single {
