+++ app/src/main/java/com/example/rastreioonibus/mapsInicio/MapsFragments.kt
@@ -1,4 +1,4 @@
-package com.example.rastreioonibus
+package com.example.rastreioonibus.mapsInicio
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -6,9 +6,17 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.rastreioonibus.databinding.LayoutMapsFragmentBinding
-import com.google.android.gms.maps.MapFragment
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.SupervisorJob
+import kotlinx.coroutines.launch
+import org.koin.android.ext.android.inject
 
 class MapsFragments: Fragment() {
+    private val viewModel: MapsViewModel by inject()
+
+    private val scope = CoroutineScope(Dispatchers.Main + SupervisorJob())
+
     private val binding: LayoutMapsFragmentBinding by lazy {
         LayoutMapsFragmentBinding.inflate(layoutInflater)
     }
@@ -23,6 +31,10 @@ class MapsFragments: Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+
+        scope.launch{
+            viewModel.autenticar(requireContext())
+        }
     }
 
     companion object{
